+++ branch-beta/sqlmath_browser.mjs
@@ -19,7 +19,6 @@ let {
 let DBTABLE_DICT = new Map();
 let DB_CHART;
 let DB_DICT = new Map();
-let DB_II = 0;
 let DB_MAIN;
 let DB_QUERY;
 let DEBOUNCE_DICT = Object.create(null);
@@ -52,9 +51,7 @@ async function dbFileAttachAsync({
 }) {
 // this function will attach database <dbData> to <db>
     let dbAttached;
-    let dbName;
-    DB_II += 1;
-    dbName = `attached_${String(DB_II).padStart(2, "0")}`;
+    let dbName = dbNameNext("attach{{ii}}", new Set(DB_DICT.keys()));
     dbAttached = await dbOpenAsync({
         dbData,
         filename: `file:${dbName}?mode=memory&cache=shared`
@@ -65,6 +62,37 @@ async function dbFileAttachAsync({
         sql: `ATTACH DATABASE '${dbAttached.filename}' AS ${dbName}`
     });
     DB_DICT.set(dbName, dbAttached);
+    // normalize order
+    DB_DICT = Array.from(DB_DICT.values());
+    DB_DICT = new Map([
+        DB_DICT.slice(0, 3),
+        DB_DICT.slice(3).sort(function (aa, bb) {
+            aa = aa.dbName;
+            bb = bb.dbName;
+            return (
+                aa < bb
+                ? -1
+                : 1
+            );
+        })
+    ].flat().map(function (db) {
+        return [
+            db.dbName, db
+        ];
+    }));
+}
+
+function dbNameNext(template, bag) {
+// this function will get next incremental name in <bag> from given <template>
+    let ii = 0;
+    let name;
+    while (true) {
+        ii += 1;
+        name = template.replace("{{ii}}", String(ii).padStart(2, "0"));
+        if (!bag.has(name)) {
+            return name;
+        }
+    }
 }
 
 async function dbTableImportAsync({
@@ -136,7 +164,7 @@ async function dbTableImportAsync({
     colList = colList.map(function (colName) {
         let colName2;
         let duplicate = 0;
-        colName = "c_" + colName.replace((
+        colName = "c_" + colName.toLowerCase().replace((
             /\W/g
         ), "_");
         while (true) {
@@ -228,8 +256,8 @@ CREATE TABLE __test2 AS
     UNION ALL SELECT 1, 2, 3
     UNION ALL SELECT 4, 5, 6;
 
-DROP TABLE IF EXISTS attached_01.__test3;
-CREATE TABLE attached_01.__test3 AS SELECT * FROM __test2;
+DROP TABLE IF EXISTS attach01.__test3;
+CREATE TABLE attach01.__test3 AS SELECT * FROM __test2;
 
 SELECT
         *,
@@ -383,15 +411,16 @@ async function demoTradebot() {
             "backtrack"
         ].map(function (dateInterval) {
             let optionDict;
+            let tableChart;
             let tableData;
-            let tableName;
             tableData = (
                 dateInterval === "1 day"
                 ? "tradebot_intraday"
                 : "tradebot_historical"
             );
-            tableName = (
-                `_{{ii}}_tradebot_historical_${dateInterval.replace(" ", "_")}`
+            tableChart = (
+                "chart._{{ii}}_tradebot_historical_"
+                + dateInterval.replace(" ", "_")
             );
             optionDict = {
                 title: (
@@ -406,6 +435,11 @@ async function demoTradebot() {
                     )
                 ),
                 xaxisTitle: "date",
+                xstep: (
+                    dateInterval === "1 day"
+                    ? 60
+                    : 1
+                ),
                 xvalueConvert: (
                     dateInterval === "1 day"
                     ? "unixepochToTimeutc"
@@ -415,9 +449,9 @@ async function demoTradebot() {
                 yvalueSuffix: " %"
             };
             return (`
--- chart - tradebot_historical
-DROP TABLE IF EXISTS chart.${tableName};
-CREATE TABLE chart.${tableName} (
+-- chart - ${tableChart} - create
+DROP TABLE IF EXISTS ${tableChart};
+CREATE TABLE ${tableChart} (
     datatype TEXT NOT NULL,
     series_index INTEGER,
     xx REAL,
@@ -426,16 +460,16 @@ CREATE TABLE chart.${tableName} (
     xx_label TEXT,
     options TEXT
 );
-INSERT INTO chart.${tableName} (datatype, options)
+INSERT INTO ${tableChart} (datatype, options)
     SELECT
         'options' AS datatype,
         '${JSON.stringify(optionDict)}' AS options;
-INSERT INTO chart.${tableName} (datatype, options, series_index, series_label)
+INSERT INTO ${tableChart} (datatype, options, series_index, series_label)
     SELECT
         'series_label' AS datatype,
         json_object(
             'isDummy', is_dummy,
-            'isHidden', sym NOT in ('__tradebot1', 'spy', 'dia', 'qqq')
+            'isHidden', sym NOT in ('01_mybot', 'spy', 'dia', 'qqq')
         ) AS options,
         rownum AS series_index,
         sym AS series_label
@@ -444,7 +478,7 @@ INSERT INTO chart.${tableName} (datatype, options, series_index, series_label)
             sym LIKE '-%' AS is_dummy,
             ROW_NUMBER() OVER (
                 ORDER BY
-                    sym = '__tradebot1' DESC,
+                    sym = '01_mybot' DESC,
                     sym = '----' DESC,
                     sym = 'spy' DESC,
                     sym = 'dia' DESC,
@@ -501,7 +535,7 @@ UPDATE __tmp1
             ydate DESC
         LIMIT 1
     );
-INSERT INTO chart.${tableName} (datatype, xx, xx_label)
+INSERT INTO ${tableChart} (datatype, xx, xx_label)
     SELECT
         'xx_label' AS datatype,
         rownum AS xx,
@@ -514,7 +548,7 @@ INSERT INTO chart.${tableName} (datatype, xx, xx_label)
         WHERE
             aa <= ydate AND ydate <= bb
     );
-INSERT INTO chart.${tableName} (datatype, series_index, xx, yy)
+INSERT INTO ${tableChart} (datatype, series_index, xx, yy)
     SELECT
         'yy_value' AS datatype,
         series_index,
@@ -527,7 +561,7 @@ INSERT INTO chart.${tableName} (datatype, series_index, xx, yy)
             SELECT
                 series_index,
                 series_label
-            FROM chart.${tableName}
+            FROM ${tableChart}
             WHERE
                 datatype = 'series_label'
         )
@@ -535,7 +569,7 @@ INSERT INTO chart.${tableName} (datatype, series_index, xx, yy)
             SELECT
                 xx,
                 xx_label
-            FROM chart.${tableName}
+            FROM ${tableChart}
             WHERE
                 datatype = 'xx_label'
         )
@@ -544,7 +578,7 @@ INSERT INTO chart.${tableName} (datatype, series_index, xx, yy)
     ON
         sym = series_label
         AND ydate = xx_label;
-UPDATE chart.${tableName}
+UPDATE ${tableChart}
     SET
             ${
                 (
@@ -572,7 +606,7 @@ UPDATE chart.${tableName}
                 ) AS rownum,
                 yy,
                 series_index
-            FROM chart.${tableName}
+            FROM ${tableChart}
             WHERE
                 datatype = 'yy_value'
                 AND yy > 0
@@ -580,7 +614,7 @@ UPDATE chart.${tableName}
         WHERE
             rownum = 1
     ) USING (series_index);
-UPDATE chart.${tableName}
+UPDATE ${tableChart}
     SET
         series_label = printf(
             '%+06.2f%% - %s%s',
@@ -600,7 +634,7 @@ UPDATE chart.${tableName}
                 ) AS rownum,
                 series_index,
                 yy AS yy_today
-            FROM chart.${tableName}
+            FROM ${tableChart}
             WHERE
                 datatype = 'yy_value'
         )
@@ -611,29 +645,29 @@ UPDATE chart.${tableName}
     WHERE
         datatype = 'series_label';
 -- chart - tradebot_historical - normalize xx to unixepoch
-DELETE FROM chart.${tableName} WHERE
+DELETE FROM ${tableChart} WHERE
     '${dateInterval}' = '1 day'
     AND datatype = 'yy_value'
     AND xx = 1;
-UPDATE chart.${tableName}
+UPDATE ${tableChart}
     SET
         xx = ${(
                 dateInterval === "1 day"
                 ? "UNIXEPOCH(tt)"
-                : "ROUND(JULIANDAY(tt))"
+                : "JULIANDAY(tt)"
             )}
     FROM (SELECT 0)
     JOIN (
         SELECT
             xx,
             xx_label AS tt
-        FROM chart.${tableName}
+        FROM ${tableChart}
         WHERE
             datatype = 'xx_label'
     ) USING (xx)
     WHERE
         datatype = 'yy_value';
-INSERT INTO chart.${tableName} (datatype, series_index, xx, yy)
+INSERT INTO ${tableChart} (datatype, series_index, xx, yy)
     SELECT
         'yy_value' AS datatype,
         series_index,
@@ -642,77 +676,150 @@ INSERT INTO chart.${tableName} (datatype, series_index, xx, yy)
     FROM (
         SELECT
             series_index
-        FROM chart.${tableName}
+        FROM ${tableChart}
         WHERE datatype = 'series_label'
     )
     JOIN (
         SELECT
             MIN(xx) AS xx0
-        FROM chart.${tableName}
+        FROM ${tableChart}
         WHERE
             datatype = 'yy_value'
     )
     WHERE
         '${dateInterval}' = '1 day';
-DELETE FROM chart.${tableName} WHERE datatype = 'xx_label';
+DELETE FROM ${tableChart} WHERE datatype = 'xx_label';
             `);
         }),
-        (`
--- chart - tradebot_performance_sector
-DROP TABLE IF EXISTS __tmp1;
-CREATE TEMP TABLE __tmp1 AS
-    SELECT
-        *
+        [
+            "sector",
+            "subsector",
+            "stock"
+        ].map(function (grouping) {
+            return [
+                "performance",
+                "holding"
+            ].map(function (ptype) {
+                let columnData;
+                let optionDict;
+                let sqlSelect;
+                let tableChart;
+                columnData = (
+                    ptype === "performance"
+                    ? "perc_gain_today"
+                    : "perc_holding"
+                );
+                optionDict = {
+                    isBarchart: true,
+                    title: `tradebot ${ptype} today by ${grouping}`,
+                    xaxisTitle: grouping,
+                    yaxisTitle: (
+                        ptype === "performance"
+                        ? "percent gain"
+                        : "percent holding"
+                    ),
+                    yvalueSuffix: " %"
+                };
+                sqlSelect = (
+                    grouping === "sector"
+                    ? (`
+SELECT
+        IIF(category LIKE 'short%', 1, grouping_index) AS series_color,
+        category LIKE '-%' AS is_dummy,
+        grouping IN ('account', 'exchange') AS is_hidden,
+        printf(
+            '%05.2f%% - %s - %s',
+            ${columnData},
+            grouping,
+            category
+        ) AS series_label,
+        ROW_NUMBER() OVER (
+            ORDER BY
+                grouping_index,
+                category != '----' DESC,
+                ${columnData} DESC
+        ) AS xx,
+        category AS xx_label,
+        ${columnData} AS yy
     FROM (
         SELECT
-            IIF(category = 'short', 1, grouping_index) AS series_color,
-            category LIKE '-%' AS is_dummy,
-            grouping IN ('account', 'exchange') AS is_hidden,
-            printf(
-                '%05.2f%% - %s - %s',
-                perc_gain_today,
-                grouping,
-                category
-            ) AS series_label,
-            ROW_NUMBER() OVER (
-                ORDER BY
-                    grouping_index,
-                    category != '----' DESC,
-                    perc_gain_today DESC
-            ) AS xx,
-            perc_gain_today AS yy
+            category,
+            grouping,
+            grouping_index,
+            ${columnData},
+            perc_holding
+        FROM tradebot_position_category
+        WHERE
+            grouping != 'subsector'
+        --
+        UNION ALL
+        --
+        SELECT
+            '----' AS category,
+            '----' AS grouping,
+            grouping_index,
+            NULL AS ${columnData},
+            NULL perc_holding
         FROM (
-            SELECT
-                category,
+            SELECT DISTINCT
                 grouping,
-                grouping_index,
-                perc_gain_today,
-                perc_holding
+                grouping_index
             FROM tradebot_position_category
-            WHERE
-                grouping != 'subsector'
-            --
-            UNION ALL
-            --
-            SELECT
-                '----' AS category,
-                '----' AS grouping,
-                grouping_index,
-                NULL AS perc_gain_today,
-                NULL perc_holding
-            FROM (
-                SELECT DISTINCT
-                    grouping,
-                    grouping_index
-                FROM tradebot_position_category
-            )
         )
     )
-    ORDER BY
-        series_color,
-        xx;
-DROP TABLE IF EXISTS chart._{{ii}}_tradebot_performance_sector;
-CREATE TABLE chart._{{ii}}_tradebot_performance_sector (
+                    `)
+                    : grouping === "subsector"
+                    ? (`
+SELECT
+        IIF(category LIKE 'short%', 1, grouping_index) AS series_color,
+        category LIKE '-%' AS is_dummy,
+        0 AS is_hidden,
+        printf('%05.2f%% - %s', ${columnData}, category) AS series_label,
+        ROW_NUMBER() OVER (
+            ORDER BY
+                grouping_index,
+                category != '----' DESC,
+                ${columnData} DESC
+        ) AS xx,
+        SUBSTR(category, INSTR(category, '____') + 4) AS xx_label,
+        ${columnData} AS yy
+    FROM (
+        SELECT
+            category,
+            grouping,
+            grouping_index,
+            ${columnData},
+            perc_holding
+        FROM tradebot_position_category
+        WHERE
+            grouping = 'subsector'
+    )
+                    `)
+                    : (`
+SELECT
+        IIF(is_short, 1, 2) AS series_color,
+        0 AS is_dummy,
+        0 AS is_hidden,
+        printf(
+            '%+.2f%% - %s - %s - %s',
+            ${columnData},
+            IIF(is_short, 'short', 'long'),
+            sym,
+            company_name
+        ) AS series_label,
+        ROW_NUMBER() OVER (ORDER BY ${columnData} DESC) AS xx,
+        sym AS xx_label,
+        ${columnData} AS yy
+    FROM tradebot_position_stock
+                    `)
+                );
+                tableChart = `chart._{{ii}}_tradebot_${grouping}_${ptype}`;
+                return (`
+-- chart - ${tableChart} - create
+DROP TABLE IF EXISTS __tmp1;
+CREATE TEMP TABLE __tmp1 AS SELECT * FROM (${sqlSelect}) ORDER BY xx;
+DROP TABLE IF EXISTS ${tableChart};
+CREATE TABLE ${tableChart} (
     datatype TEXT NOT NULL,
     series_index INTEGER,
     xx REAL,
@@ -721,17 +828,11 @@ CREATE TABLE chart._{{ii}}_tradebot_performance_sector (
     xx_label TEXT,
     options TEXT
 );
-INSERT INTO chart._{{ii}}_tradebot_performance_sector (datatype, options)
+INSERT INTO ${tableChart} (datatype, options)
     SELECT
         'options' AS datatype,
-        '{
-            "isBarchart": true,
-            "title": "tradebot performance today by sector",
-            "xaxisTitle": "category",
-            "yaxisTitle": "percent gain",
-            "yvalueSuffix": " %"
-        }' AS options;
-INSERT INTO chart._{{ii}}_tradebot_performance_sector (
+        '${JSON.stringify(optionDict)}' AS options;
+INSERT INTO ${tableChart} (
     datatype,
     options,
     series_index,
@@ -747,162 +848,13 @@ INSERT INTO chart._{{ii}}_tradebot_performance_sector (
         xx AS series_index,
         series_label
     FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_performance_sector (
-    datatype,
-    series_index,
-    xx,
-    yy
-)
-    SELECT
-        'yy_value' AS datatype,
-        xx AS series_index,
-        xx,
-        yy
-    FROM __tmp1;
-        `),
-        (`
--- chart - tradebot_performance_subsector
-DROP TABLE IF EXISTS __tmp1;
-CREATE TEMP TABLE __tmp1 AS
-    SELECT
-        *
-    FROM (
-        SELECT
-            IIF(category = 'short____short', 1, grouping_index) AS series_color,
-            printf(
-                '%05.2f%% - %s',
-                perc_gain_today,
-                category
-            ) AS series_label,
-            ROW_NUMBER() OVER (
-                ORDER BY
-                    grouping_index,
-                    category != '----' DESC,
-                    perc_gain_today DESC
-            ) AS xx,
-            perc_gain_today AS yy
-        FROM (
-            SELECT
-                category,
-                grouping,
-                grouping_index,
-                perc_gain_today,
-                perc_holding
-            FROM tradebot_position_category
-            WHERE
-                grouping = 'subsector'
-        )
-    )
-    ORDER BY
-        series_color,
-        xx;
-DROP TABLE IF EXISTS chart._{{ii}}_tradebot_performance_subsector;
-CREATE TABLE chart._{{ii}}_tradebot_performance_subsector (
-    datatype TEXT NOT NULL,
-    series_index INTEGER,
-    xx REAL,
-    yy REAL,
-    series_label REAL,
-    xx_label TEXT,
-    options TEXT
-);
-INSERT INTO chart._{{ii}}_tradebot_performance_subsector (datatype, options)
-    SELECT
-        'options' AS datatype,
-        '{
-            "isBarchart": true,
-            "title": "tradebot performance today by subsector",
-            "xaxisTitle": "category",
-            "yaxisTitle": "percent gain",
-            "yvalueSuffix": " %"
-        }' AS options;
-INSERT INTO chart._{{ii}}_tradebot_performance_subsector (
-    datatype,
-    options,
-    series_index,
-    series_label
-)
-    SELECT
-        'series_label' AS datatype,
-        json_object('seriesColor', series_color) AS options,
-        xx AS series_index,
-        series_label
-    FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_performance_subsector (
-    datatype,
-    series_index,
-    xx,
-    yy
-)
-    SELECT
-        'yy_value' AS datatype,
-        xx AS series_index,
-        xx,
-        yy
-    FROM __tmp1;
-        `),
-        (`
--- chart - tradebot_performance_stock
-DROP TABLE IF EXISTS __tmp1;
-CREATE TEMP TABLE __tmp1 AS
-    SELECT
-        *
-    FROM (
-        SELECT
-            IIF(is_short, 1, 2) AS series_color,
-            printf(
-                '%+.2f%% - %s - %s - %s',
-                perc_gain_today,
-                IIF(is_short, 'short', 'long'),
-                sym,
-                company_name
-            ) AS series_label,
-            ROW_NUMBER() OVER (ORDER BY perc_gain_today DESC) xx,
-            sym AS xx_label,
-            perc_gain_today AS yy
-        FROM tradebot_position_stock
-    )
-    ORDER BY
-        xx;
-DROP TABLE IF EXISTS chart._{{ii}}_tradebot_performance_stock;
-CREATE TABLE chart._{{ii}}_tradebot_performance_stock (
-    datatype TEXT NOT NULL,
-    series_index INTEGER,
-    xx REAL,
-    yy REAL,
-    series_label REAL,
-    xx_label TEXT,
-    options TEXT
-);
-INSERT INTO chart._{{ii}}_tradebot_performance_stock (datatype, options)
-    SELECT
-        'options' AS datatype,
-        '{
-            "isBarchart": true,
-            "title": "tradebot performance today by stock",
-            "xaxisTitle": "stock",
-            "yaxisTitle": "percent gain",
-            "yvalueSuffix": " %"
-        }' AS options;
-INSERT INTO chart._{{ii}}_tradebot_performance_stock (
-    datatype,
-    options,
-    series_index,
-    series_label
-)
-    SELECT
-        'series_label' AS datatype,
-        json_object('seriesColor', series_color) AS options,
-        xx AS series_index,
-        series_label
-    FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_performance_stock (datatype, xx, xx_label)
+INSERT INTO ${tableChart} (datatype, xx, xx_label)
     SELECT
         'xx_label' AS datatype,
         xx,
         xx_label
     FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_performance_stock (
+INSERT INTO ${tableChart} (
     datatype,
     series_index,
     xx,
@@ -914,265 +866,9 @@ INSERT INTO chart._{{ii}}_tradebot_performance_stock (
         xx,
         yy
     FROM __tmp1;
-        `),
-        (`
--- chart - tradebot_position_sector
-DROP TABLE IF EXISTS __tmp1;
-CREATE TEMP TABLE __tmp1 AS
-    SELECT
-        *
-    FROM (
-        SELECT
-            IIF(category = 'short', 1, grouping_index) AS series_color,
-            category LIKE '-%' AS is_dummy,
-            grouping IN ('account', 'exchange') AS is_hidden,
-            printf(
-                '%05.2f%% - %s - %s',
-                perc_holding,
-                grouping,
-                category
-            ) AS series_label,
-            ROW_NUMBER() OVER (
-                ORDER BY
-                    grouping_index,
-                    category != '----' DESC,
-                    perc_holding DESC
-            ) AS xx,
-            perc_holding AS yy
-        FROM (
-            SELECT
-                category,
-                grouping,
-                grouping_index,
-                perc_gain_today,
-                perc_holding
-            FROM tradebot_position_category
-            WHERE
-                grouping != 'subsector'
-            --
-            UNION ALL
-            --
-            SELECT
-                '----' AS category,
-                '----' AS grouping,
-                grouping_index,
-                NULL AS perc_gain_today,
-                NULL perc_holding
-            FROM (
-                SELECT DISTINCT
-                    grouping,
-                    grouping_index
-                FROM tradebot_position_category
-            )
-        )
-    )
-    ORDER BY
-        series_color,
-        xx;
-DROP TABLE IF EXISTS chart._{{ii}}_tradebot_position_sector;
-CREATE TABLE chart._{{ii}}_tradebot_position_sector (
-    datatype TEXT NOT NULL,
-    series_index INTEGER,
-    xx REAL,
-    yy REAL,
-    series_label REAL,
-    xx_label TEXT,
-    options TEXT
-);
-INSERT INTO chart._{{ii}}_tradebot_position_sector (datatype, options)
-    SELECT
-        'options' AS datatype,
-        '{
-            "isBarchart": true,
-            "title": "tradebot position today by sector",
-            "xaxisTitle": "category",
-            "yaxisTitle": "percent holding",
-            "yvalueSuffix": " %"
-        }' AS options;
-INSERT INTO chart._{{ii}}_tradebot_position_sector (
-    datatype,
-    options,
-    series_index,
-    series_label
-)
-    SELECT
-        'series_label' AS datatype,
-        json_object(
-            'isDummy', is_dummy,
-            'isHidden', is_hidden,
-            'seriesColor', series_color
-        ) AS options,
-        xx AS series_index,
-        series_label
-    FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_position_sector (
-    datatype,
-    series_index,
-    xx,
-    yy
-)
-    SELECT
-        'yy_value' AS datatype,
-        xx AS series_index,
-        xx,
-        yy
-    FROM __tmp1;
-        `),
-        (`
--- chart - tradebot_position_subsector
-DROP TABLE IF EXISTS __tmp1;
-CREATE TEMP TABLE __tmp1 AS
-    SELECT
-        *
-    FROM (
-        SELECT
-            IIF(category = 'short____short', 1, grouping_index) AS series_color,
-            printf(
-                '%05.2f%% - %s',
-                perc_holding,
-                category
-            ) AS series_label,
-            ROW_NUMBER() OVER (
-                ORDER BY
-                    grouping_index,
-                    category != '----' DESC,
-                    perc_holding DESC
-            ) AS xx,
-            perc_holding AS yy
-        FROM (
-            SELECT
-                category,
-                grouping,
-                grouping_index,
-                perc_gain_today,
-                perc_holding
-            FROM tradebot_position_category
-            WHERE
-                grouping = 'subsector'
-        )
-    )
-    ORDER BY
-        series_color,
-        xx;
-DROP TABLE IF EXISTS chart._{{ii}}_tradebot_position_subsector;
-CREATE TABLE chart._{{ii}}_tradebot_position_subsector (
-    datatype TEXT NOT NULL,
-    series_index INTEGER,
-    xx REAL,
-    yy REAL,
-    series_label REAL,
-    xx_label TEXT,
-    options TEXT
-);
-INSERT INTO chart._{{ii}}_tradebot_position_subsector (datatype, options)
-    SELECT
-        'options' AS datatype,
-        '{
-            "isBarchart": true,
-            "title": "tradebot position today by subsector",
-            "xaxisTitle": "category",
-            "yaxisTitle": "percent holding",
-            "yvalueSuffix": " %"
-        }' AS options;
-INSERT INTO chart._{{ii}}_tradebot_position_subsector (
-    datatype,
-    options,
-    series_index,
-    series_label
-)
-    SELECT
-        'series_label' AS datatype,
-        json_object('seriesColor', series_color) AS options,
-        xx AS series_index,
-        series_label
-    FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_position_subsector (
-    datatype,
-    series_index,
-    xx,
-    yy
-)
-    SELECT
-        'yy_value' AS datatype,
-        xx AS series_index,
-        xx,
-        yy
-    FROM __tmp1;
-        `),
-        (`
--- chart - tradebot_position_stock
-DROP TABLE IF EXISTS __tmp1;
-CREATE TEMP TABLE __tmp1 AS
-    SELECT
-        *
-    FROM (
-        SELECT
-            IIF(is_short, 1, 2) AS series_color,
-            printf(
-                '%05.2f%% - %s - %s - %s',
-                perc_holding,
-                IIF(is_short, 'short', 'long'),
-                sym,
-                company_name
-            ) AS series_label,
-            ROW_NUMBER() OVER (ORDER BY perc_holding DESC) xx,
-            sym AS xx_label,
-            perc_holding AS yy
-        FROM tradebot_position_stock
-    )
-    ORDER BY
-        xx;
-DROP TABLE IF EXISTS chart._{{ii}}_tradebot_position_stock;
-CREATE TABLE chart._{{ii}}_tradebot_position_stock (
-    datatype TEXT NOT NULL,
-    series_index INTEGER,
-    xx REAL,
-    yy REAL,
-    series_label REAL,
-    xx_label TEXT,
-    options TEXT
-);
-INSERT INTO chart._{{ii}}_tradebot_position_stock (datatype, options)
-    SELECT
-        'options' AS datatype,
-        '{
-            "isBarchart": true,
-            "title": "tradebot position today by stock",
-            "xaxisTitle": "stock",
-            "yaxisTitle": "percent holding",
-            "yvalueSuffix": " %"
-        }' AS options;
-INSERT INTO chart._{{ii}}_tradebot_position_stock (
-    datatype,
-    options,
-    series_index,
-    series_label
-)
-    SELECT
-        'series_label' AS datatype,
-        json_object('seriesColor', series_color) AS options,
-        xx AS series_index,
-        series_label
-    FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_position_stock (datatype, xx, xx_label)
-    SELECT
-        'xx_label' AS datatype,
-        xx,
-        xx_label
-    FROM __tmp1;
-INSERT INTO chart._{{ii}}_tradebot_position_stock (
-    datatype,
-    series_index,
-    xx,
-    yy
-)
-    SELECT
-        'yy_value' AS datatype,
-        xx AS series_index,
-        xx,
-        yy
-    FROM __tmp1;
-        `),
+                `);
+            });
+        }),
         (`
 -- chart - tradebot_buysell_history
 DROP TABLE IF EXISTS __tmp1;
@@ -1245,7 +941,7 @@ INSERT INTO chart._{{ii}}_tradebot_buysell_history (
     SELECT
         'series_label' AS datatype,
         json_object('seriesColor', series_color) AS options,
-        xx AS series_index,
+        __tmp1.rowid AS series_index,
         series_label
     FROM __tmp1;
 INSERT INTO chart._{{ii}}_tradebot_buysell_history (
@@ -1256,12 +952,12 @@ INSERT INTO chart._{{ii}}_tradebot_buysell_history (
 )
     SELECT
         'yy_value' AS datatype,
-        xx AS series_index,
+        __tmp1.rowid AS series_index,
         xx,
         yy
     FROM __tmp1;
         `)
-    ].flat().map(function (sql, ii) {
+    ].flat().flat().map(function (sql, ii) {
         return sql.trim().replace((
             /\{\{ii\}\}/g
         ), String(ii + 1).padStart(2, "0"));
@@ -1372,7 +1068,7 @@ async function init() {
         ["body", "click", onContextmenu],
         ["body", "contextmenu", onContextmenu],
         [UI_FILE_OPEN, "change", onDbAction],
-        [document, "keyup", onKeyUp],
+        [document, "keydown", onKeyDown],
         [window, "hashchange", uitableInitWithinView],
         [window, "resize", onResize]
     ].forEach(function ([
@@ -1862,6 +1558,12 @@ RENAME TO
         }
         uiFadeIn(UI_CRUD);
         UI_CRUD.style.zIndex = 1;
+        // ergonomy - auto-focus first input
+        Array.from(UI_CRUD.querySelectorAll(
+            ".modalContent input"
+        )).every(function (elem) {
+            elem.focus();
+        });
         return;
     }
     // slow actions that require loading
@@ -1915,11 +1617,10 @@ RENAME TO
                 ? "csv"
                 : "json"
             ),
-            tableName: (
-                action === "dbtableImportCsv"
-                ? `__file_csv_${Date.now()}`
-                : `__file_json_${Date.now()}`
-            ),
+            tableName: dbNameNext(
+                "[__file{{ii}}]",
+                new Set(DB_MAIN.dbtableList.keys())
+            ).slice(1, -1),
             textData: (
                 await target.files[0].text()
             )
@@ -2146,14 +1847,8 @@ async function onDbcrudExec({
     UI_CRUD.style.zIndex = -1;
 }
 
-function onKeyUp(evt) {
+function onKeyDown(evt) {
 // this function will handle event keyup
-    if (!evt.modeDebounce) {
-        debounce("onKeyUp", onKeyUp, Object.assign(evt, {
-            modeDebounce: true
-        }));
-        return;
-    }
     switch (evt.key) {
     case "Escape":
         // close error-modal
@@ -2163,8 +1858,21 @@ function onKeyUp(evt) {
         return;
     }
     switch ((evt.ctrlKey || evt.metaKey) && evt.key) {
+    // database - execute
     case "Enter":
         onDbExec({});
+        return;
+    // database - open
+    case "o":
+        evt.preventDefault();
+        evt.stopPropagation();
+        document.querySelector("button[data-action='dbOpen']").click();
+        return;
+    // database - save
+    case "s":
+        evt.preventDefault();
+        evt.stopPropagation();
+        document.querySelector("button[data-action='dbExportMain']").click();
         return;
     }
 }
@@ -2582,9 +2290,7 @@ async function uiRenderDb() {
         }
     });
     // DBTABLE_DICT - sync
-    await Promise.all(Array.from(
-        DB_DICT.values()
-    ).map(async function (db) {
+    await Promise.all(Array.from(DB_DICT.values()).map(async function (db) {
         let baton;
         let {
             dbName,
@@ -2594,7 +2300,7 @@ async function uiRenderDb() {
         } = db;
         let dbtableList;
         let tmp;
-        db.dbtableList = [];
+        db.dbtableList = new Map();
         dbtableList = noop(
             await dbExecAsync({
                 db,
@@ -2606,14 +2312,14 @@ SELECT * FROM sqlite_schema WHERE type = 'table' ORDER BY tbl_name;
         if (!dbtableList) {
             return;
         }
-        dbtableList = dbtableList.map(function ({
+        dbtableList = new Map(dbtableList.map(function ({
             colList = [],
             dbtableFullname,
             rowCount = 0,
             rowList0,
             sql,
             tbl_name
-        }) {
+        }, ii) {
             dbtableFullname = dbtableFullname || `${dbName}.[${tbl_name}]`;
             dbtableIi += 1;
             baton = {
@@ -2624,6 +2330,7 @@ SELECT * FROM sqlite_schema WHERE type = 'table' ORDER BY tbl_name;
                 dbtableIi,
                 dbtableName: `[${tbl_name}]`,
                 hashtag: `dbtable_${String(dbtableIi).padStart(2, "0")}`,
+                ii,
                 isDbchart,
                 isDbmain,
                 isDbquery,
@@ -2635,8 +2342,10 @@ SELECT * FROM sqlite_schema WHERE type = 'table' ORDER BY tbl_name;
                 title: `table ${dbtableFullname}`
             };
             DBTABLE_DICT.set(baton.hashtag, baton);
-            return baton;
-        });
+            return [
+                baton.dbtableName, baton
+            ];
+        }));
         tmp = "";
         dbtableList.forEach(function ({
             dbtableName,
@@ -2722,8 +2431,9 @@ SELECT COUNT(*) AS rowcount FROM ${dbtableName};
             dbtableFullname,
             dbtableName,
             hashtag,
+            ii,
             rowCount
-        }, ii) {
+        }) {
             html += `<a class="tocElemA"`;
             html += ` data-hashtag="${hashtag}"`;
             html += ` href="#${hashtag}"`;
@@ -2808,6 +2518,19 @@ async function uichartCreate(baton) {
     let uichart = await dbExecAsync({
         db,
         sql: (`
+-- table - __chart_options1 - insert
+DROP TABLE IF EXISTS __chart_options1;
+CREATE TEMP TABLE __chart_options1 AS
+    SELECT
+        json(options) AS options,
+        CASTREALORZERO(options->>'$.xstep') AS xstep,
+        CASTREALORZERO(1.0 / options->>'$.xstep') AS xstep_inv,
+        CASTREALORZERO(options->>'$.ystep') AS ystep,
+        CASTREALORZERO(1.0 / options->>'$.ystep') AS ystep_inv
+    FROM (
+        SELECT options FROM ${dbtableName} WHERE datatype = 'options' LIMIT 1
+    );
+
 -- table - __chart_series_xy1 - insert
 DROP TABLE IF EXISTS __chart_series_xy1;
 CREATE TEMP TABLE __chart_series_xy1 AS
@@ -2832,10 +2555,18 @@ CREATE TEMP TABLE __chart_series_xy1 AS
                 series_index,
                 xx,
                 yy
-            FROM ${dbtableName}
-            WHERE
-                datatype = 'yy_value'
-                AND xx IS NOT NULL
+            FROM (
+                SELECT
+                    ${dbtableName}.rowid,
+                    series_index,
+                    IIF(xstep_inv, ROUND(xstep_inv * xx), xx) AS xx,
+                    IIF(ystep_inv, ROUND(ystep_inv * yy), yy) AS yy
+                FROM ${dbtableName}
+                JOIN __chart_options1
+                WHERE
+                    datatype = 'yy_value'
+                    AND xx IS NOT NULL
+            )
             ORDER BY
                 series_index,
                 xx
@@ -2859,6 +2590,7 @@ CREATE TEMP TABLE __chart_series_maxmin1 AS
     GROUP BY
         series_index;
 
+-- table - __chart_options1 - select - options
 SELECT
         json_set(
             options,
@@ -2870,7 +2602,7 @@ SELECT
             '$.ydataMax', ydataMax,
             '$.ydataMin', ydataMin
         ) AS options
-    FROM (SELECT options FROM ${dbtableName} WHERE datatype = 'options' LIMIT 1)
+    FROM __chart_options1
     JOIN (
         SELECT
             json_group_array(xx_label) AS xlabelList
@@ -3182,8 +2914,9 @@ SELECT
                     : tickx
                 ),
                 prefix: uichart[xOrY + "valuePrefix"],
+                step: uichart[xOrY + "step"],
                 suffix: uichart[xOrY + "valueSuffix"]
-            });
+            }).slice(0, 16);
             // number already formatted
             if (typeof num !== "number") {
                 return num;
@@ -3244,9 +2977,13 @@ SELECT
         modeTick,
         num,
         prefix,
+        step,
         suffix
     }) {
     // this function will format <num>
+        if (step && typeof num === "number") {
+            num *= step;
+        }
         switch (convert) {
         case "juliandayToDate":
             num = new Date((num - 2440587.5) * 86400 * 1000);
@@ -3358,12 +3095,14 @@ SELECT
             convert: uichart.xvalueConvert,
             num: xlabelList[xval - 1] ?? xval,
             prefix: uichart.xvaluePrefix,
+            step: uichart.xstep,
             suffix: uichart.xvalueSuffix
         });
         ylabel = numberFormat({
             convert: uichart.yvalueConvert,
             num: yval,
             prefix: uichart.yvaluePrefix,
+            step: uichart.ystep,
             suffix: uichart.yvalueSuffix
         });
         // update elemTooltipText
@@ -4645,25 +4384,3 @@ function waitAsync(timeout) {
 
 // init
 window.addEventListener("load", init);
-
-export {
-    DB_MAIN,
-    UI_EDITOR,
-    assertOrThrow,
-    dbCloseAsync,
-    dbExecAsync,
-    dbFileAttachAsync,
-    dbFileExportAsync,
-    dbFileImportAsync,
-    dbOpenAsync,
-    debugInline,
-    domDivCreate,
-    jsonHtmlSafe,
-    noop,
-    onDbExec,
-    sqlmathWebworkerInit,
-    stringHtmlSafe,
-    uiFadeIn,
-    uiFadeOut,
-    uiRenderDb
-};